Factor out Intelephense
