+++ extensions/php/src/language_servers.rs
@@ -1,0 +1,3 @@
+mod intelephense;
+
+pub use intelephense::*;
