+++ extensions/php/src/php.rs
@@ -1,64 +1,18 @@
-use std::{env, fs};
+mod language_servers;
+
+use std::env;
+
 use zed_extension_api::{self as zed, LanguageServerId, Result};
 
-const SERVER_PATH: &str = "node_modules/intelephense/lib/intelephense.js";
-const PACKAGE_NAME: &str = "intelephense";
+use crate::language_servers::Intelephense;
 
 struct PhpExtension {
-    did_find_server: bool,
-}
-
-impl PhpExtension {
-    fn server_exists(&self) -> bool {
-        fs::metadata(SERVER_PATH).map_or(false, |stat| stat.is_file())
-    }
-
-    fn server_script_path(&mut self, language_server_id: &LanguageServerId) -> Result<String> {
-        let server_exists = self.server_exists();
-        if self.did_find_server && server_exists {
-            return Ok(SERVER_PATH.to_string());
-        }
-
-        zed::set_language_server_installation_status(
-            &language_server_id,
-            &zed::LanguageServerInstallationStatus::CheckingForUpdate,
-        );
-        let version = zed::npm_package_latest_version(PACKAGE_NAME)?;
-
-        if !server_exists
-            || zed::npm_package_installed_version(PACKAGE_NAME)?.as_ref() != Some(&version)
-        {
-            zed::set_language_server_installation_status(
-                &language_server_id,
-                &zed::LanguageServerInstallationStatus::Downloading,
-            );
-            let result = zed::npm_install_package(PACKAGE_NAME, &version);
-            match result {
-                Ok(()) => {
-                    if !self.server_exists() {
-                        Err(format!(
-                            "installed package '{PACKAGE_NAME}' did not contain expected path '{SERVER_PATH}'",
-                        ))?;
-                    }
-                }
-                Err(error) => {
-                    if !self.server_exists() {
-                        Err(error)?;
-                    }
-                }
-            }
-        }
-
-        self.did_find_server = true;
-        Ok(SERVER_PATH.to_string())
-    }
+    intelephense: Option<Intelephense>,
 }
 
 impl zed::Extension for PhpExtension {
     fn new() -> Self {
-        Self {
-            did_find_server: false,
-        }
+        Self { intelephense: None }
     }
 
     fn language_server_command(
@@ -66,19 +20,26 @@ impl zed::Extension for PhpExtension {
         language_server_id: &LanguageServerId,
         _worktree: &zed::Worktree,
     ) -> Result<zed::Command> {
-        let server_path = self.server_script_path(language_server_id)?;
-        Ok(zed::Command {
-            command: zed::node_binary_path()?,
-            args: vec![
-                env::current_dir()
-                    .unwrap()
-                    .join(&server_path)
-                    .to_string_lossy()
-                    .to_string(),
-                "--stdio".to_string(),
-            ],
-            env: Default::default(),
-        })
+        match language_server_id.as_ref() {
+            Intelephense::LANGUAGE_SERVER_ID => {
+                let intelephense = self.intelephense.get_or_insert_with(|| Intelephense::new());
+
+                let server_path = intelephense.server_script_path(language_server_id)?;
+                Ok(zed::Command {
+                    command: zed::node_binary_path()?,
+                    args: vec![
+                        env::current_dir()
+                            .unwrap()
+                            .join(&server_path)
+                            .to_string_lossy()
+                            .to_string(),
+                        "--stdio".to_string(),
+                    ],
+                    env: Default::default(),
+                })
+            }
+            language_server_id => Err(format!("unknown language server: {language_server_id}")),
+        }
     }
 }
